build fixes
